shell: Save the width and height known by the shell surface.
Instead of using the surface size, save the last known shell surface
size, and use that one to emit on configure, when the surface is
returning from maximized or fullscreen state.

The surface size can't be saved because it will have already changed
when the surface is transitioning to fullscreen or maximized.